Enable building Mac/Windows specific solution projects
Rearrange the directory structure of the repository to make it easier
to selectively build platform-specific code.

The common/ directory contains any code that can be built and run on all
platforms. The windows/ (and mac/, linux/, etc as needed) directory
contains code that can only be built and/or run on Windows.

Change the main .NET solution to be built for either Windows or Mac -
building only those projects which are required for Windows/Mac.
This is done via solution configurations, of which there are now four:

 - Debug
 - Release
 - WindowsDebug
 - WindowsRelease

The Windows* configurations will build all the common class libraries
(netstandard2.0), the Windows helpers (net461) and the CLI head project
(netcoreapp2.1).

The Debug/Release configurations will build only the common libraries
(netstandard2.0) and the CLI head project (netcoreapp2.1). This is
suitable for building on Mac & Linux (and Windows without the helpers).