Use the global alloc/free methods for native interop
Replace usages of the COM allocator (`Marshal.AllocCoTaskMem` and
`Marshal.FreeCoTaskMem`) in the credential stores code with the
global allocator (`Marshal.AllocHGlobal` and `Marshal.FreeHGlobal`).

Since there is no COM heap on Mac using this allocator was not the
correct thing to do. On Windows the CredRead/Write/Delete calls are
Win32 API calls which don't use COM either.

Use the global allocator for WindowsCredMgr native code