Add ability to override host provider from configuration
Check the Git configuration for a specified host provider for the
current repository remote URI.

We look for the `credential.authority` configuration entry (and any URL
scoped variants) for backwards compatibility with older GCMs, as well as
the new `credential.provider` entry.

Also introduce a new setting `credential.allowWindowsAuth` to allow users
who used the `basic` authority to disable WIA detection of the new generic
provider.