Add INI serializer
Add a serializer (and deserializer) for INI files, specifically the
Git variant which includes 'contexts' or 'scopes' on section headers.

INI files can be used to store cached or stateful information for the
host providers.

The rationale for this format other others (say JSON or binary) is
consistency with Git configuration.