Return to building 64-bit in NetCore and NetFx on Windows
Now that we've moved the Microsoft authentication helper in-process we
no longer have a need to force all entry executables to be the same
bit-width.

The GitHub authentication helper is still .NET Framework (as we're
targeting .NET Core 2.x still) and using 32-bit WinForms. However that
helper does not make use of the LibGit2 native bits (by virtue of
not instantiating the `CommandContext` type).