Add GCM_INTERACTIVE/credential.interactive setting
Add a setting to disable all user interaction. By setting
`GCM_INTERACTIVE` or `credential.interactive` to a 'falsey' value, GCM
Core will now fail and return an error if user interaction is required.

This is useful in headless and unattended environments, such as build
servers, where is is preferable to fail than it is to hang waiting for
input from a non-existent user.

The setting also existed in the previous GCM for Windows, but its
behaviour has been slightly modified to treat 'always' values as 'auto'.
See more in the documentation and the code.

The default value is `true` / permit interaction.