winui: make SecurePasswordBox notify all changes
Update the `SecurePasswordBox` control to notify all listeners of
changes to the password value. Previously we only notified on the first
character typed into the box.

We clone the current password, modify it, and then replace the previous
instance (triggering an `INotifyPropertyChanged::PropertyChanged`
event). With a binding delay, we will batch updates of multiple
keystrokes rather than one at a time.