azrepos: use organizations authority for MSA accounts
Use the /organizations authority for MSA accounts with Azure
DevOps/Repos. This is because we're using MSA pass-through, an internal
Microsoft mechanism to support both MSA and 'work' (AAD) accounts with
the same auth stacks.

You should be able to use /common, but this doesn't work. At the same
time we're using ADAL Obj-C on macOS rather than MSAL.NET like we do on
Windows, and ADAL speaks to the "v1" AAD endpoints, which don't know the
/organizations tenant :(

For macOS we need to fudge the authority _back_ to /common for MSA
accounts.