azrepos: workaround org@ user hackery
Now that we support multiple user accounts for each host/service/remote,
we have hit an interesting issue with Azure Repos.

With the introduction of the dev.azure.com-style URLs for Azure Repos
there was an unfortunate hack or workaround invented to add the AzDevOps org name
to the userinfo part of the remote URL, for example: org@dev.azure.com/org/blah.

Since GCM for Windows (and older versions of GCM Core that initially followed
the same model) always uses the value "PersonalAccessToken" for the username field
when storing credentials it was free to ignore the user part of the input.

The problem now is that since we support multiple user accounts, and will perform
an exact match against the credential (with user) if a username is specified in
the remote URL, we never find the credential we now store (we now always store
with the actual, real users' UPN).

To workaround this workaround (yuck) we ignore the username IF AND ONLY IF the
host is dev.azure.com, and return the first matching dev.azure.com/org
credential. The upshot of this is that dev.azure.com-style URLs do NOT
support multiple users OR full paths, however vs.com-style URLs will.