wincredmgr: recover the correct service name
Recover the correct service name from the target name.

Since the target name may contain a userinfo component (for example
https://alice_domain.com@example.com/path), and the only place we store
the service name is in the target name, we need to strip out any
userinfo component.

We do this by looking for the "://" and the first '@' character before
the first '/', which act as the start and end of the userinfo component.