gitconfig: avoid deadlock on large amounts of config data
As per the Remarks section in
https://docs.microsoft.com/en-us/dotnet/api/system.diagnostics.processstartinfo.redirectstandardoutput
we should first read the output until the end before waiting for the
`git config` process to finish, to avoid deadlocks.

This indeed helps in the situation of this developer whose config is so
large that `git config -l` produces 25 kilobyte worth of output.
Apparently, this fills up some buffer that seems to be around 8kB and
then `git config -l` waits for its output to be consumed, and before
this patch, the consumer (i.e. GCM Core) would wait for `git config -l`
to exit already.