windows: split Windows installer in to user/system
Split the Windows installer into two versions: a single-user installer,
and a system-wide installer.

The old installer installed in Program Files (x86) and set the system
Git configuration, which is now what the system-flavoured installer
does.

The new single-user installer places files in %LOCALAPPDATA% and updates
the ~/.gitconfig (global) user Git configuration instead.

The benefit of the single-user installer is that it does not require an
administrator to install it, and doesn't affect other users of the
system.