http: treat empty cred proxy as system default creds
To enable using the default system credentials for authenticated proxies
we must set the `WebProxy.UseDefaultCredentials` property to `true`.
(For example, with Windows Integrated Authentication-based proxies -
NTLM or Kerberos.)

Currently we set this if no username and password information is
specified at all in the proxy configuration. However, Git itself when
presented with this configuration will prompt the user for a
username/password (which isn't required).

Specifying an empty string username/password in the proxy configuration
will stop Git from prompting for them, and allow cURL to auto-negotiate
proxy auth. However, GCM reads that same configuration as having a
username/password based proxy with the empty string for both values.

We change GCM to use the system default credentials when configuring a
proxy if there is no userinfo OR if BOTH the username and password
values are the empty string.

Example proxy config value: http://:@proxy.example.com