env: enumerate the PATH envar to locate programs on Windows
Enumerate the PATH environment variable to manually locate
executables/programs rather than shell out to where.exe on Windows
locate these for us. The where.exe utility first checks the current
working directory before the PATH for the executable, which we do not
want for various security reasons.

The which utility on UNIX/POSIX systems does _NOT_ include the current
working directory in the search; only the PATH, which is exactly what we
want - leave this in place on those platforms.

Additionally for locating `git(.exe)` we first check in the
`GIT_EXEC_PATH` directory for the executable, if that environment
variable was set.