config: match Git config mixed-sensitively
Git configuration entry keys are neither case sensitive, nor
in-sensitive. They are mixed sensitive.

The section and property names are NOT sensitive, but the 'scope' or
inner component of the full name IS sensitive.

For example, `credential.example.com.provider` is the same as
`CREDENTIAL.example.com.PROVIDER`, but `credential.EXAMPLE.COM.provider`
is NOT.

We introduce a new StringComparer type that performs insensitive
comparison on the section and property components, and sensitive
comparison on the scope component.