github: update Windows GitHub UI to add PAT support
Update the WPF-based Windows GUI for GitHub authentication prompts to
support the third authentication mode "PAT".

Also make some changes to how the dialog window is sized to scale with
the size of the content (mins and maxes included), which may now vary in
the extreme from all three auth options, to just one.