azrepos: add authority cache for AzDevOps orgs
Add a cache of AAD authorities for Azure DevOps organisations, using an
INI file.

To avoid making a HEAD request to the Azure DevOps instance each time we
want to perform AAD/MSA authentication in order to determine the
authority, we first check for a cached value and use that.

If there is an auth failure then we remove any cached authority for that
AzDevOps org in case the reason for failure was a token from the wrong
issuing authority. This can happen if an AzDevOps org changes its
backing authority from MSA/AAD or between AAD tenants.