azrepos: use remote URL user if present in JWT mode
When running in AAD Access token (JWT) mode, if the remote URL contains
a username _and_ is using the classic org.visualstudio.com domain-name,
then attempt to acquire an AAT AT for that user. Otherwise use the user
manager's bindings.

We can only offer this for vs.com-style URLs and _not_ dev.azure.com
URLs because of the decision by Azure Repos to default include the
organisation name as the user-info part of generated clone URLs to
workaround the Git credential design of not including the full path
(required to know which tenant an org is backed by).

For dev.azure.com URLs we can only rely on the user manager and explicit
bindings.