xui: add a dispatcher class
Add a dispatcher type for use with Avalonia. This dispatcher is
different from the built-in Avalonia `Dispatcher` type.

On macOS the underlying platform requires that the main NSRunLoop be
run on the **entry** thread of the process (that is, thread 0).
This is unlike on Windows where all that is required is an "STA" thread.

Because of this we need to "reserve" thread 0 (the main entry thread)
for running the Avalonia infrastructure, and kick off a new thread to
run the rest of the "normal" .NET application logic (such as
command-line handling).

The dispatcher allows queuing of work to be run on the initialised
thread, until the loop is terminated. We can use this dispatcher and
model to start the Avalonia app/runloops at a defered time, after main
app startup.

The Avalonia `Dispatcher` type is then used to push more work on to that
inner message pump/runloop after the Avalonia application has started.