github: fix bug in x-UI cmdline parsing
The System.CommandLine command line binding requires the parameter names
match the option names.

During the rebasing of the x-platform UI work, the argument name for the
`--oauth` option was left as `browser` which means the OAuth browser
flow is never shown, even though the base GCM program is passing the
right argument/options.