linux: add option for GCM_GPG_PATH environment variable
If GCM_GPG_PATH is specified, it is assumed to be the path to the GPG
executable used by `pass` and will throw an exception if it is not found.
If GCM_GPG_PATH is _not_ specified, GCM Core will attempt to find the
correct GPG executable by mimicking the behavior (currently) used by `pass`
(that is, first searching for `gpg2`, then falling back on `gpg` if `gpg2` is
not found).