osx: dynamically skip keychain test if invalid
Dynamically skip the macOS Keychain read/write/delete test if the
Keychain is in a "strange" state.

There is an unknown issue that the keychain can sometimes get itself
in where all API calls result in an errSecAuthFailed error. The only
solution seems to be a machine restart; not possible in CI!

The problem has plagued others who are calling the same Keychain APIs
from C# such as the MSAL.NET team - they don't know either. It might
have something to do with the code signing signature of the binary
(our collective best theory).

It's probably only diagnosable at this point by Apple, but we don't
have a reliable way to reproduce the problem.