git: add `GitVersion` type and `Version` property to `IGit`
The `Version` property (calculated only once, on first access) has type
`GitVersion`, made up of the pre-custom components of a Git version (i.e.,
before a `.windows` or `.vfs`). This version is currently only used in
`GitConfiguration` in order to determine whether to use the now-deprecated
canonical type constraint flags (for backwards-compatibility purposes)
when accessing a config value.

Signed-off-by: Victoria Dye <vdye@github.com>