registry: speed up auto-detect and make more robust
Speed up the auto-detection mechanism in the HostProviderRegistry by
setting a short(er) timeout on the probing network call (2 seconds).
The default value from the framework is otherwise 100 seconds(!).

We introduce a setting for the user to be able to configure this timeout
value with the granularity of 1 millisecond.

Another problem that occured with this network call was crashes due to
TLS configuration issues. The network call was not wrapped in a
try-catch block which meant that the entire process would crash if there
was a problem. This isn't ideal for what is supposed to be a "best
effort" feature. Here we wrap the HEAD call in a try-catch and display
an appropriate warning message to the user.

For completeness, add documentation that explains the auto-detection
feature, how to configure the timeout, and how to disable this network
call.