README: add information on removal from  MSFT Ubuntu feeds
As part of the migration of GCM  out of the microsoft org, we removed
packages from the Microsoft Hirsute/Bionic feeds. Adding a note to the
README informing users that these packages are no longer available and
that installing the debian package from the releases page is the preferred
mode at this time.