gcm: create shims & links for older GCMC exec name
Create shim scripts for callers of the older "GCM Core" executable name.
The scripts will warn users that the name of the executable has changed
and should be updated, and then forward invocation to the GCM
executable.

We also create symlinks in the Debian package and macOS package to these
shim scripts. This is to allow external tools that have not updated to use
the new name to continue to work.

Symlinks are the natural solution here and are used on UNIX platforms.

However, since they are not widely supported on Windows, we don't use
shims. Given that GCM is not on the %PATH% in Windows by default, we hope
that our friends in the Git for Windows can resolve this similarly in their
distribution.