dependencies: update Microsoft.NET.Test.Sdk and Avalonia
The upcoming specification of a macOS runtime led to the need to update the
Microsoft.NET.Test.Sdk package. The old version, when combined with this
specification, caused dependent packages to be downgraded, which resulted
in NuGet errors. Upgrading to the latest version resolved this issue.
Additionally, Avalonia has been upgraded to the latest version, which was
necessitated by the upcoming changes to support osx-arm64.