.azure-pipelines: restore and specify mac project
The specification of a macOS runtime resulted in the need to change how
GCM compiles macOS projects in its release pipeline. Previously, GCM built
the entire solution in the compile phase of this pipeline. However, when
`--runtime` is explicitly specified, `dotnet` also requires explicit
specification of the project(s) to build (i.e. building an entire solution
is no longer permitted). This change also entailed an explicit restore,
without which the Installer.Mac project build failed.