Start using different distribution.xml files for different architectures
Instead of using a universal distribution.xml file, they are split up, that way if an Apple Silicon user goes to install the Intel version by mistake, they will know when they go to install, rather than having to actually execute the binary to get prompted about installing Rosetta.

When universal binaries are able to be done in .NET Core, this change will be able to be undone.