bitbucket: show on-prem URL in UI and use better help links
Update the Bitbucket credential UI to show the URL of the Bitbucket
Server or Data Center instance, if we're talking to them.

Also fix up the help links to point to a host-relative location; not
always to Bitbucket Cloud pages. Note that the OAuth prompt will never
been shown for Server/DC since 2FA is not supported there.