windows: generate app manifests for all exe projects
Automatically generate the correct application manifest for Windows
executable projects that we produce. The manifests are required on
Windows systems newer than Windows 8 to correctly access versioning
APIs that will otherwise always report "Windows 8", even on later
versions.

The manifest we generate states we are compatible with all versions of
Windows from 7 to 11 (inclusive).