Use /usr/bin/env which instead of /usr/bin/which
It may cause trouble to hard-code `/usr/bin/which` anywhere, so for now, why not `/usr/bin/env` to resolve which? See https://github.com/GitCredentialManager/git-credential-manager/pull/717 for example.

(In the long run it may cause trouble to use `which` at all since it's not POSIX standard and Debian is thinking of removing it. But then you could probably just ensure another package is installed first.)