workflows: auto-generate maintainer away issue
Add a new workflow to create issues that will notify GCM users and
contributors of when maintainers are away and the repo is not being
actively monitored. This workflow is controlled by a `workflow_dispatch`
trigger that requires both a start date and end date for maintainer
absence before running.