avn: workaround a macOS window focus bug
For some reason on macOS windows that are activated are not appearing on
top, and get lost behind other application windows. The platform
implementation on macOS is seemingly correctly calling the
`makeKeyAndOrderFront` native windowing API, but it's not being made
"key and front".

Setting the "topmost" flag true (and then false again) brings our
windows to the front, which is what we want.