release: fix Windows archives
When we transitioned our release workflow to GitHub actions, we started
using Powershell's `Compress-Archive` cmdlet to zip our Windows binaries.
However, this unfortunately does not play nicely with `unzip` in WSL.
Since we're publishing the archives from an Ubuntu machine, switch to
creating the archives with the `zip` command in the default bash shell to
correct the issue.