Updates RFC links to authoritative URL
The current URLs use an outdated source.

The rfc-editor.org site is stated to be the authoritative source, see aaragraph 3 on https://www.ietf.org/standards/rfcs/.

The redirection used at the existing URL cause the later versions of the link checker to fail for the instances in the documentation. However, instances in the source code comments have also been updated.