release: publish gcm as dotnet tool
Add a new `DotnetTool` project containing the needed files to publish GCM
as a cross-platform dotnet tool. Additionally, update GCM to use helper
dlls if exes are not found in the Application directory. At present, dlls
are only used in the dotnet tool scenario to minimize the package size and
to make it cross-platform compatible.