dotnet tool: rename package
Change the name and author of the dotnet tool package from
GitCredentialManager.Cli to git-credential-manager. The reason for this
change is that we were able to transfer ownership of this package (which
was previously owned by JetBrains) to the git-credential-manager
organization on NuGet.org.  There should not be any customer impact
related to this change, as the GitCredentialManager.Cli package had yet
to be published.