gcm: create symlinks and warn message for gcmcore exec
Create symlinks and shim/copy-executables for the original exec
"git-credential-manager-core(.exe)" name for consumers who have not
updated to the new name.

We detect if the consumer is launching us via the "-core" symlink or
executable shim by consulting the platform-native APIs to get the
original "argv[0]". All of the .NET APIs sadly don't give us the real
"argv[0]", so we need to use native APIs...

If we detect use of the old name, print a warning that the user should
update their configuration, and a help link for more information.