dotnet-tool: add GCM icon to tool nuget package
Add the GCM icon to the NuGet package used for the dotnet tool.
The image is copied to the root of the package under a new `images`
directory (/images). The tool payload continues to be copied under the
required `/tool/net6.0/any` directory.

Modify the pack-tool.sh script to keep all build output under the /out
directory of the repository (which is in .gitignore). Output from
running pack-tool.sh now looks like this:

.
└── out
    └── shared
        └── DotnetTool
            └── nupkg
                └── Debug
                    ├── git-credential-manager.$VERSION.nupkg
                    ├── images       # icon
                    ├── payload      # dotnet publish output
                    └── payload.sym  # pdbs

The contents of the NuGet package looks like this:

.
├── [Content_Types].xml
├── _rels
├── git-credential-manager.nuspec
├── images
│   └── icon.png          # copied from the output images/ dir
├── package
│   └── services
│       └── metadata/...
└── tools
    └── net6.0
        └── any/...       # copied from the output payload/ dir