Update markdownlint-cli2-action to 7.0.0
This required fixing some newly found lint issues.

Additionally some issues are in files (workflow templates) that are structured in such a way that do not easily allow correction, so I have adjusted the linter configuration to ignore these files for now. They could instead be corrected in a later commit.