diagnose: always use IHttpClientFactory for diagnose
The Network diagnostic currently makes two passes of the tests, one
using a 'raw' `HttpClient` and another using one created by the product
`HttpClientFactory`.

However, we've not actually been using the factory-built client in the
tests! Furthermore, the raw HTTP client will not exercise test correctly
if the machine is behind a corporate proxy, so let's just drop that
option.