docs: Update README docs links
Update README docs links to point at release branch. The goal of this
change is to avoid confusing users when we merge major changes (e.g. the
executable rename, the dotnet tool install option, etc.) to main but
have not yet released them. With this update, the documentation should
align with the latest release.