fs: resolve symlinks and fix bugs in *FileSystem impls
There is a systematic bug in all the IFileSystem implementations in that
we were incorrectly calling `Path.GetFileName` rather than
`Path.GetFullPath`, meaning we're only comparing the file _names_ of two
paths for equality! Fix this.

For Mac and Linux we also handle symlinks (or firmlinks) by now using
the .NET APIs to resolve links. On .NET Framework we don't do anything.