apppath: introduce ICommandContext.InstallationDirectory
Introduce a new property on the ICommandContext that represents the GCM
installation directory, separate from the application executable path.

The app executable is not always found in the installation directory if
the app exe path is a symlink (or we're a .NET tool where the AppHost is
outside of the tool store).