argv0: fix implementation of getting native argv[0]
Our previous implementation of getting the native argv[0] was flawed on
Linux and Mac. The Mac `_NSArgv` and Linux `/proc/self/cmdline`
function/proc-file return the raw argv[0] which may not be an absolute
or rooted path, but we expected this to be the same.

Additionally, the logic of appending the CWD to the argv[0] value when
the latter wasn't rooted was also just wrong!

Update to now use `_NSExecutablePath` on macOS, which always returns the
full absolute path, and add correct handling of unrooted argv[0] on
Linux.

Furthermore, on Linux, if we are unable to handle the value from
/proc/self/cmdline, we fall back to resolving the /proc/self/exe symlink
instead.