app: fix installation dir calculation
On Windows, when invoked as a partial name (e.g., `manager`) our command
line argv[0] is just a file name, not a full file path! We incorrectly
assumed that `Environment.GetCommandLineArgs()[0]` was always absolute!

We should instead use `AppContext.BaseDirectory` that, in all
circumstances tested, all on TFMs and OSs, all publish options, and as a
.NET tool... returns the expected full, absolute path to the
installation directory.