app: fix bug in oldname warning on Windows
Fix a bug in the old-name warning detection that only presents on
Windows when bundled with Git for Windows, and when configured using the
shorthand config `credential.helper=manager-core`.
In this scenario, our argv[0] is missing the ".exe" extension (that's
actually present in the file name of course) because ... mingw reasons.

Instead of matching "git-credential-manager-core.exe" on Windows, just
strip ".exe" if present on Windows, and always match
"git-credential-manager-core" on all platforms.