env: ensure we don't return empty PATH vars on Windows
Ensure that we don't return empty `PATH` values on Windows when
splitting. It's common that on Windows the `PATH` can be inadvertently
modified such that an empty value is included, and this can lead
to weird resolution of executables using the current directory.
As it stands, no existing codepaths would trigger this today when
invoked by Git for Windows, but let's make sure this never happens in
the future by mistake!