app: fix Windows app-path resolution
Fix another bug in how we compute the Application Executable path on
Windows. It some cases we can still get a relative/non-absolute path to
the entry executable. If we get a filename/non-path from the native API
calls then return null and fallback to using process/module image
resolution instead. For .NET Framework on Windows, this is OK.
We will want to revisit this on .NET (Core) on Windows, if and when we
get around to that...