release: force xz compression for Debian package
The GCM Debian package currently fails to install on Debian with the
following error:

error: archive 'gcm-linux_amd64.2.0.886.deb' uses unknown compression for
member 'control.tar.zst', giving up

It appears that the version of dpkg that ships with Debian does not
support zstd compression [1]. Enforcing xz compression resolves the issue.

This also provides an opportunity to clean up some unused variables in
pack.sh and ensure we check the architecture is found before attempting to
use the ARCH variable.

[1]: https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=892664