docs: update git credential cache platforms
GCM's documentation states that git's credential cache is supported on
Windows. Unfortunately, due to lack of Unix socket support on Windows
versions prior to Windows 10, this feature is not currently supported on
Windows, so this change removes it from the list of platforms on which
this credstore can be used. See [1] for more details.

[1]: https://github.com/git-for-windows/git/issues/3892