trace2: add sid
A key component of Git's TRACE2 tracing system is the session id (sid).
This identifies the process instance to allow all events emitted by it to
be identified.

We check to see if a parent sid (i.e. from a Git process) exists. If so,
we separate this sid from the GCM sid (a GUID) using a "/". If there is no
parent sid, we simply generate a GUID for the GCM sid and use that.

The above also requires addition of a new SetEnvironmentVariable() method
in EnvironmentBase.cs to set the GCM-specific SID variable.