trace2: detect event/normal formats
Add the infrastructure to detect whether a user has enabled the TRACE2
event format or normal format targets. This implementation has been
designed to be extended to include the perf format target in a future
series.

Additionally it involved some refactoring/cleanup to set the Application
Path and InstallationDirectory in the CommandContext, rather than in
GCM/UI helper Program.cs files.