trace2: refactor to prepare for child process-related events
The TRACE2 child start and child exit events will require some refactoring
around the way GCM handles processes to provide finer-grained control and
the ability to capture process-related information.

In preparation for this, we need to do some re-organization of the Trace2
code. With this change we:

1. Move TRACE2 tracing startup into Program.cs for GCM and the helper
exes. This allows us to start tracing earlier and capture more events.
2. Add an Initialize() method to Trace2.cs to detect whether Trace2
tracing is enabled and set long-lived variables. This also sets a new
_initialized property that allows us to determine whether Trace2 tracing
has been initialized. Trace2 tracing is disabled until _initialized is set
to true by the Initialize() method.
3. Change the name of TryParseSettings() to InitializeWriters() for
improved clarity.
4. Pass the CommandContext to Trace2 to use its Streams and Environment.