.net tool: clarify install instructions
Because we use dotnet publish (which requires a specific framework) to
create the payload of our .NET tool package, the package is tied to the
framework specified. This means that it will only work with the
corresponding .NET SDK for that framework.

In light of this, update the .NET tool install instructions to specify
that users must download the latest .NET 6.0 SDK as a prerequisite for
installing the package, since that is the framework for which we
currently publish. This will also need to be updated whenever we upgrade
to a new Target Framework.

Additionally, remove Windows as a supported .NET tool platform. Since we
currently target .NET Framework rather than .NET, the package cannot
actually be installed on Windows.