trace2: redirect standard error
When a new process is created, standard error can optionally be
redirected. This is useful in situations where the output is verbose and
should be suppressed rather than output to the console, for example.

However, this can cause an issue with TRACE2 deadlocks, as Git writes to
standard error and GCM does nothing. This change corrects this by not
redirecting standard error for Git-related processes (which also extends
to places such as Environment.cs, given Git's reliance on envars).

Finally, there are some places where we should still redirect standard
error for the reason mentioned above. For these scenarios, comments have
been added to help readers understand the purposeful nature of these
redirects.