dict: add support for read/writing multidictionaries
Add support for reading and writing multi-dictionaries in Git's
credential helper I/O format. Multi-dictionaries differ from normal
dictionaries by being key:value relation 1:N.

Git prepends "[]" to the end of keys that should be treated as lists or
having mutliple values. An empty value in the list has the effect of
cleaning/resetting the values in the list so far.

The first multiple-valued key is expected to be 'wwwauth[]' from this
patch series:

https://lore.kernel.org/git/pull.1352.v11.git.1677518420.gitgitgadget@gmail.com/