trace2: add performance format target
Add the performance format target for TRACE2 tracing. This change
includes a few notable components:

1. A new GetMessage method to ensure message strings written to normal and
performance targets are not duplicated.
2. A new PerformanceFormatComponent class to track sizing of optional
properties of performance format messages.
3. A BuildSpan method that Trace2Message children use to create
correctly-sized "spans" for optional properties of performance format
messages. A span is a piece of the message beginning with a pipe (|) and
ending just before the next pipe or the end of the message.
4. A BuildTimeSpan method that adjusts spans for long/short times. In
future, when we have events that include performance format string span
components, we will need to add an additional method to handle them.