trace2: add string extension for snake casing
The TRACE2 convention is for multi-word event names to be written in snake
case (all lower case with words separated by underscores). While up until
this point we've had single-word event names, this series will add the
multi-word child start and child exit events. In anticipation of that,
this change adds a string extension that will be used in later patches to
write multi-word enum values in snake case.