trace2: create sid manager class
Create static class to manage session ids (sids). This allows us to
immediately set up GCM's sid in Program.cs to ensure Git processes that
are executed before Trace2 tracing is set up have the correct parent sid.
We also use it to give helper exes unique child sids.

Although this will also execute when TRACE2 tracing is not enabled, it
should not be an issue - it will simply set the (unused)
GIT_TRACE2_PARENT_SID varible for the process to a GUID.