trace2: add file writer
Our current strategy of using a Stream to write to files for Trace2
tracing is not working - Git processes are overwriting GCM's data and vice
versa.

To resolve this, switch from using the Trace2StreamWriter to write to
files to a dedicated Trace2FileWriter class. This also includes a refactor
to add a parent base class for Trace2Writers for shared functaionality
between the Trace2StreamWriter and Trace2FileWriter.