trace2: add process depth calculation
The performance format target requires recording of process depth. We use
the number of forward slashes in the process's sid as a convenient way to
determine this. Because the sid and the depth are process-related
concepts, this change also moves sid/depth related properties into
ProcessManager.cs and removes the SidManager.cs class.