env: avoid virtual member call in ctor; all refresh
We had a virtual member call in the constructor of the various
IEnvironment implementations, which is a code smell.

Instead, lazily populate the `_variables` cache dictionary, and add an
explict `Refresh` method to refresh the cache. We call `Refresh` in the
`SetEnvironmentVariable` method for convenience.

For testing we simply pipe through the pre-computed variables in the
`internal` constructor.