trace2: add error tracing statements
Add TRACE2 error tracing statements throughout GCM codebase. Note that
this is a best effort - there are certain places (most notably certain
static and unsafe methods) where statements were purposefully not added
because it is either not safe or was deemed to much lift/churn to do so.

Note that there are some instances in which a "parameterized" message is
passed to TRACE2. This is used only when PII information could be passed
into the message (e.g. through a path or remote URI) and should be
redacted for collection purposes.

Finally, there are certain instances in which we write an error with no
exception thrown. These align with the places where we are currently
writing to GCM Trace without throwing an error for the purposes of parity.