trace2: add region infrastructure
Add supporting infrastructure for regions. This includes the normal
`Write` methods in Trace2 and corresponding Trace2 messages with a few
notable distinctions:

1. There is an abstract parent RegionMessage that inherits from
Trace2Message. The RegionEnterMessage and RegionLeave messages inherit
from this parent (rather than Trace2Message) to maximize reuse of shared
fields.
2. We create a new Region class that implements DisposableObject and call
it using a CreateRegion method in Trace2. This ensures that we call
WriteRegionLeave before a Region object is disposed - a key requirement of
Git's Trace2 API for regions.  Additionally, it allows us to manage
getting the executing thread name and region timings within the class,
thus removing this burden from consumers.

An additional note is that the region_enter and region_leave events will
always have the same line number with this model. This was deemed to be
acceptable, however, given that knowing where a region begins executing
provides enough information for inspection of the region if it is needed.