trace2: capture thread name
With the upcoming introduction of regions, we will need to capture thread
names. This change adds a new static BuildThreadName() method that defines
a thread name by:

1. Determining whether the executing thread is from the .NET thread pool.
If so, the naming convention is thread_pool_<thread_id>.
2. If the thread is not from the .NET thread pool, check and see if it's
named. If it is, the name is returned. If not, an empty string is
returned.

Child processes and the upcoming implementation of regions use this
method. However, in places where we know the name of the executing thread
for sure (e.g. Program.cs for the main GCM exe and helper exes), we simply
use that name.