trace2: add regions to OAuth
Add region tracing to key methods in OAuth2Client. This was deemed a good
starting point for regions due to the server interaction required to
obtain auth codes, device codes, and tokens. Additional regions can and
should be added to additional sections of the code that are deemed
performance-critical in the future.