settings: cache Git config values across process
When enumerating all Git configuration values in `Settings`, we
currently cache a copy of all values for the lifetime of the
`GetSettingValues()` method. This saves us some calls to Git, but we can
do better.

Pull-up the cache of Git configuration entries to the lifetime of the
`Settings` object, meaning all calls to `GetSettingValues()` over the
lifetime of the process can benefit from the cache! The savings here can
be significant as each `ISettings` property typically results in a call
to `GetSettingValues()`.