msauth: use new Windows broker based on MSALRuntime
Use the new Windows broker which is based on the MSALRuntime; an export
wrapper around a native, cross-platform MSAL library.

In this new set up, we drop the `.Desktop` package in favour of the
`.Broker` package that also means we drop the WebView2Loader.dll, which
we didn't make use of anyway. There are a few new binaries to be
distrubuted in the new model, including a P/Invoke layer, IdentityModel
abstractions library, and the native msalruntime_x86.dll.

Note that GCM still only support x86 on Windows, and only supports
broker use on Windows. For this reason we don't bother adding the broker
package on non-.NET Framework builds to keep the sizes on Mac/Linux to a
minimium.