msauth: always set the parent window for MSAL
Always set the parent window handle for MSAL on Windows.

Previously we only set this if provided a handle by the user/config. Now
however we must always try and provide a handle because using the new
MSALRuntime-based Windows broker means we must do so - MSAL no longer
provides us with a 'dummy' handle to use.

Use the parent console window handle, as recommended by the MSAL docs:
https://github.com/AzureAD/microsoft-authentication-library-for-dotnet/wiki/wam#parent-window-handles