gitlab: simplify terminal/interaction checks
Simplify where we check for user interaction and/or terminal prompts
during authentication, to only need to check at the start of each
`GetAuthenticationAsync` call.

Note that is is a behaviour change - we now require user interaction to
be allowed to perform authentication even with only 1 mode available.
This is technically more 'correct' because when the user disables
interactivity, we should never require interaction to continue. Although
we are not asking the user to select a mode of authentication, we still
either need them to enter a username/password, or respond to a browser
OAuth flow.