bitbucket: lazily init the HttpClient for OAuthRegistry
Lazily acquire the HttpClient instance from the factory inside of the
OAuth2ClientRegistry. There is a relatively large cost to creating a
new HttpClient instance (especially the first; due to config reading)
that is being paid each time we just register the entire Bitbucket host
provider.

Whilst we are here also cleanup some of the styling/field naming in
this class to be consistent with the rest of GCM.