platformutils: use .NET Environment.OSVersion on CoreCLR
Since .NET 5 we can now use the `Environment.OSVersion` property to
lookup the real OS version for macOS and Windows[1].

For .NET Framework (still used for our releases on Windows) we must
continue to use the Win32 API `RtlGetVersion` to avoid any Windows
compatibility nonsense.

We continue to use `uname` on Linux.

[1] https://learn.microsoft.com/en-us/dotnet/core/compatibility/core-libraries/5.0/environment-osversion-returns-correct-version