release: decouple .NET tool
Move deployment of .NET tool to nuget.org out of release workflow. With
these changes, the .NET tool is published only after a release has been
published (rather than with the creation of the release). This emulates
what we currently do with Homebrew and ensures we only release the .NET
tool when we are confident enough in a release's stability to publish it.