system.text.json: bitbucket rest api custom converter
The Bitbucket authority returns the non-standard 'scopes' property with
token endpoint results. With Json.NET, it was possible to handle this
differentiation from the 'scope' property returned by other providers
using inheritance.

However, while this somewhat works in System.Text.Json, there is an issue
with overwriting the value with the final property. So, for example, if
'scopes' is the final property in the JSON returned from Bitbucket's OAuth
token endpoint, then the scopes property is correctly used in
deserialization. However, if 'scope' is the final property, this value
overwrites the scope value. For this reason, we use a custom converter to
explicitly deserialize Bitbucket OAuth responses.

For simplicity in working with System.Text.Json, we also store the
'expires_in' in a long rather than a TimeSpan.