github: fix bug passing username hint to cred lookup
In adding support for multiple-user accounts for GitHub we introduced a
bug that would mean we ignored a username-hint provided by Git/remote
URL when looking up an existing credential. The result is that we'd
return _any_ credential rather than specifically one for the desired
account!

Fix this by always passing the input username hint when first evaluating
which user account to select.