github: use 127.0.0.1 redirect over localhost
Use an IPv4 loopback redirect URL instead of the `localhost` name. This
is in accordance with the recommendation in the OAuth spec[1] and also
GitHub's documentation[2].

Note that this change depends on an update to the Git Credential Manager
OAuth application on GitHub to add the "http://127.0.0.1/" redirect (with
a trailing slash!). We will be strictly adding the new URL, and keep the
older localhost-based redirect URL untouched for older clients.

[1] https://datatracker.ietf.org/doc/html/rfc8252#section-7.3
[2] https://docs.github.com/en/apps/oauth-apps/building-oauth-apps/authorizing-oauth-apps#loopback-redirect-urls