install from source: designate archlinux as unsupported
Arch Linux was incorrectly added to the install-from-source script and
accompanying validation as a .NET-supported distribution, when this is not
actually the case. A list of supported distributions can be found here:

https://learn.microsoft.com/en-us/dotnet/core/install/linux

Instead of removing unsupported distros, we plan to print out a message
notifying users that, while they can be installed, they are not officially
validated or supported by the GCM project.