msauth: workaround MSAL.NET issue with MSA-PT accounts
When we have a Microsoft Account (MSA) in the cache and attempt to do a
silent authentication, if we're an MSA-PT app we need to specify the
special MSA transfer tenant ID to make sure we get the a token silently,
correctly. See the issue [1] in the MSAL repo for more information.

[1] https://github.com/AzureAD/microsoft-authentication-library-for-dotnet/issues/3077