msauth: fix MSA silent auth with MSA-PT
When using Microsoft Account Passthrough (MSA-PT) we need to use the
special "transfer" or "Microsoft services" tenant ID rather than the
actual MSA tenant ID when doing silent authentication.

This is a shortcoming in the MSAL library that we will need to
workaround until this issue [1] can be fixed in MSAL itself.

Modify the silent auth method such that if we are using MSA-PT, and
the `IAccount` object has the MSA tenant ID, we need explicitly set
the tenant ID to the transfer tenant ID.

Whilst we are in here, also add an extra `catch` block around the silent
auth code to capture any unexpected exceptions and log them.

[1] https://github.com/AzureAD/microsoft-authentication-library-for-dotnet/issues/3077