docs: clarify WAM + MSA connected account behaviour
When you sign-in with an MSA using WAM you are asked if you wish to use
this account "everywhere on your device", or just "Microsoft apps only".
The former will connect your local user account to an MSA, and the
latter will not.

Clarify the docs around this, including an annotated image of the
specific diaglog.